pyproject.toml
[project]
name = "shopping-cart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Memoona Zahur", email = "memoonazahur137@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.302",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.1.0",
]

[project.scripts]
shopping-cart = "shopping_cart:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
